requirements.txt
streamlit>=1.30.0
rembg>=2.0.50
pillow>=10.0.0
pydub>=0.25.1
numpy>=1.24.0
onnxruntime>=1.15.0
opencv-python>=4.8.0
scipy>=1.10.0
static-ffmpeg>=2.5.0
